pyproject.toml
[tool.poetry]
name = "challenge-mle"
version = "0.1.0"
description = ""
authors = ["DanielOspina23 <betospina23@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = ">=0.86.0,<0.87.0"
pydantic = ">=1.10.2,<1.11.0"
uvicorn = ">=0.15.0,<0.16.0"
numpy = ">=1.22.4,<1.23.0"
pandas = ">=1.3.5,<1.4.0"
scikit-learn = ">=1.3.0,<1.4.0"
matplotlib = "3.8.0"
seaborn = "0.12.2"
xgboost = ">=2.1.2,<2.2.0"
google-cloud-storage = ">=2.18.2,<2.19.0"
google-cloud-bigquery = ">=3.26.0,<3.27.0"
redis = "^5.2.0"
typing-extensions = "^4.12.2"
flake8 = "^7.1.1"
coverage = ">=5.5,<6.0"
pytest = ">=6.2.5,<6.3.0"
pytest-cov = ">=2.12.1,<2.13.0"
mockito = ">=1.2.2,<1.3.0"
anyio = "3.4.0"
google-cloud-logging = "^3.11.3"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
